+++ apps/extension/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "window.ai",
+  "name": "extension",
   "displayName": "window.ai",
   "version": "0.0.2",
   "description": "Use your own AI models on the web",
