+++ apps/web/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "windowai.io",
+  "name": "web",
   "repository": "https://github.com/alexanderatallah/window.ai",
   "license": "MIT",
   "private": true,
@@ -25,6 +25,7 @@
   "devDependencies": {
     "@types/node": "^17.0.45",
     "@types/react": "latest",
+    "window-ai": "workspace:^",
     "autoprefixer": "^10.4.14",
     "eslint": "^8.36.0",
     "eslint-config-next": "^12.3.4",
