+++ apps/web/window.ai.d.ts
@@ -1,49 +1,49 @@
-declare namespace ai {
-  type Input =
-    | {
-        prompt: string;
-      }
-    | {
-        messages: ChatMessage[];
-      };
+// declare namespace ai {
+//   type Input =
+//     | {
+//         prompt: string;
+//       }
+//     | {
+//         messages: ChatMessage[];
+//       };
 
-  type Output =
-    | {
-        text: string;
-      }
-    | {
-        message: ChatMessage;
-      };
+//   type Output =
+//     | {
+//         text: string;
+//       }
+//     | {
+//         message: ChatMessage;
+//       };
 
-  type LLM =
-    | "openai/gpt3.5"
-    | "together/gpt-neoxt-20B"
-    | "cohere/xlarge"
-    | "local";
+//   type LLM =
+//     | "openai/gpt3.5"
+//     | "together/gpt-neoxt-20B"
+//     | "cohere/xlarge"
+//     | "local";
 
-  interface CompletionOptions {
-    onStreamResult?: (result: Output | null, error: string | null) => unknown;
-    temperature?: number;
-    maxTokens?: number;
-    stopSequences?: string[];
-    model?: LLM;
-  }
+//   interface CompletionOptions {
+//     onStreamResult?: (result: Output | null, error: string | null) => unknown;
+//     temperature?: number;
+//     maxTokens?: number;
+//     stopSequences?: string[];
+//     model?: LLM;
+//   }
 
-  interface ChatMessage {
-    content: string;
-    role: string;
-  }
+//   interface ChatMessage {
+//     content: string;
+//     role: string;
+//   }
 
-  function getCurrentModel(): Promise<LLM>;
+//   function getCurrentModel(): Promise<LLM>;
 
-  function getCompletion(
-    input: Input,
-    options?: CompletionOptions
-  ): Promise<Output>;
-}
+//   function getCompletion(
+//     input: Input,
+//     options?: CompletionOptions
+//   ): Promise<Output>;
+// }
 
-declare global {
-  interface Window {
-    ai: typeof ai;
-  }
-}
+// declare global {
+//   interface Window {
+//     ai: typeof ai;
+//   }
+// }
