--- packages/test/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "test",
-  "scripts": {
-    "build": "echo building test",
-    "lint": "echo linting test",
-    "dev": "echo dev text"
-  },
-  "dependencies": {
-  },
-  "devDependencies": {
-  }
-}
